bug di lino

+++ WordCount/src/main/java/it/unisa/di/soa2019/indexing/IndexingReducer.java
@@ -1,5 +1,6 @@
 package it.unisa.di.soa2019.indexing;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.MapWritable;
 import org.apache.hadoop.io.Text;
@@ -17,7 +18,7 @@ public class IndexingReducer extends Reducer<Text, StringWritable, Text, MapWrit
 //        log.info("" + key + " values " + values.toString() + " on " + this.hashCode());
         MapWritable map = new <Text, IntWritable>MapWritable();
         for (StringWritable sw : values) {
-            map.put(new Text(sw.getKey()), new IntWritable(Integer.parseInt(sw.getVal())));
+            map.put(new Text(sw.getKey()), new DoubleWritable(Double.parseDouble(sw.getVal())));
         }
 
         context.write(key, map);
