+++ WordCount/src/main/java/it/unisa/di/soa2019/similarity/SimilarityMapper.java
@@ -1,10 +1,7 @@
 package it.unisa.di.soa2019.similarity;
 
 import it.unisa.di.soa2019.indexing.StringWritable;
-import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.MapWritable;
-import org.apache.hadoop.io.Text;
-import org.apache.hadoop.io.Writable;
+import org.apache.hadoop.io.*;
 import org.apache.hadoop.mapreduce.Mapper;
 import org.slf4j.LoggerFactory;
 
@@ -12,7 +9,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SimilarityMapper extends Mapper<Text, MapWritable, StringWritable, IntWritable> {
+public class SimilarityMapper extends Mapper<Text, MapWritable, StringWritable, DoubleWritable> {
     private static final org.slf4j.Logger log = LoggerFactory.getLogger(SimilarityMapper.class);
 
     @Override
@@ -26,12 +23,12 @@ public class SimilarityMapper extends Mapper<Text, MapWritable, StringWritable, 
         for (int i = 0; i < keys.size(); i++) {
             for (int j = i + 1; j < keys.size(); j++) {
                 Writable groupIID = keys.get(i);
-                IntWritable frequencyI = (IntWritable) value.get(keys.get(i));
+                DoubleWritable frequencyI = (DoubleWritable) value.get(keys.get(i));
 
                 Writable groupJID = keys.get(j);
-                IntWritable frequencyJ = (IntWritable) value.get(keys.get(j));
+                DoubleWritable frequencyJ = (DoubleWritable) value.get(keys.get(j));
 
-                IntWritable w = new IntWritable(frequencyI.get() * frequencyJ.get());
+                DoubleWritable w = new DoubleWritable(frequencyI.get() * frequencyJ.get());
                 StringWritable sp = new StringWritable(groupIID.toString(), groupJID.toString());
                 context.write(sp, w);
             }
